+++ README.md
@@ -7,7 +7,7 @@
 - side case에 대한 예외처리
 - 인수 테스트와 단위 테스트 작성
 - API 문서를 작성하고 문서화를 위한 테스트 작성
-- 페이지 연동
+- 페이지 연동 (Join.js 구현)
 
 ### 기능목록
 1. 회원가입
