+++ src/main/java/wooteco/subway/web/member/LoginMemberController.java
@@ -8,9 +8,6 @@ import wooteco.subway.service.member.dto.LoginRequest;
 import wooteco.subway.service.member.dto.MemberResponse;
 import wooteco.subway.service.member.dto.TokenResponse;
 
-import javax.servlet.http.HttpSession;
-import java.util.Map;
-
 @RestController
 public class LoginMemberController {
     private MemberService memberService;
@@ -25,20 +22,7 @@ public class LoginMemberController {
         return ResponseEntity.ok().body(new TokenResponse(token, "bearer"));
     }
 
-    @PostMapping("/login")
-    public ResponseEntity login(@RequestParam Map<String, String> paramMap, HttpSession session) {
-        String email = paramMap.get("email");
-        String password = paramMap.get("password");
-        if (!memberService.loginWithForm(email, password)) {
-            throw new InvalidAuthenticationException("올바르지 않은 이메일과 비밀번호 입력");
-        }
-
-        session.setAttribute("loginMemberEmail", email);
-
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping({"/me/basic", "/me/session", "/me/bearer"})
+    @GetMapping("/me/bearer")
     public ResponseEntity<MemberResponse> getMemberOfMineBasic(@LoginMember Member member) {
         return ResponseEntity.ok().body(MemberResponse.of(member));
     }
