+++ src/main/java/wooteco/subway/web/member/login/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web.member;
+package wooteco.subway.web.member.login;
 
 import org.apache.logging.log4j.util.Strings;
 import org.springframework.stereotype.Component;
