+++ src/main/java/wooteco/subway/web/member/login/BearerAuthInterceptor.java
@@ -1,10 +1,11 @@
-package wooteco.subway.web.member.interceptor;
+package wooteco.subway.web.member.login;
 
 import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.ModelAndView;
 import wooteco.subway.infra.JwtTokenProvider;
-import wooteco.subway.web.member.AuthorizationExtractor;
+import wooteco.subway.web.member.InvalidAuthenticationException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -22,15 +23,21 @@ public class BearerAuthInterceptor implements HandlerInterceptor {
     @Override
     public boolean preHandle(HttpServletRequest request,
                              HttpServletResponse response, Object handler) {
-        // TODO: Authorization 헤더를 통해 Bearer 값을 추출 (authExtractor.extract() 메서드 활용)
+        String token = authExtractor.extract(request, "Bearer");
+        validateToken(token);
 
-        // TODO: 추출한 토큰값의 유효성 검사 (jwtTokenProvider.validateToken() 메서드 활용)
-
-        // TODO: 추출한 토큰값에서 email 정보 추출 (jwtTokenProvider.getSubject() 메서드 활용)
-        String email = "";
-
+        String email = jwtTokenProvider.getSubject(token);
         request.setAttribute("loginMemberEmail", email);
         return true;
+    }
+
+    private void validateToken(String token) {
+        if (StringUtils.isEmpty(token)) {
+            throw new InvalidAuthenticationException("비정상적인 접근입니다.");
+        }
+        if(!jwtTokenProvider.validateToken(token)){
+            throw new InvalidAuthenticationException("비정상적인 접근입니다.");
+        }
     }
 
     @Override
