+++ src/main/java/wooteco/subway/web/member/login/LoginMemberMethodArgumentResolver.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web.member;
+package wooteco.subway.web.member.login;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.core.MethodParameter;
@@ -9,6 +9,8 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import wooteco.subway.domain.member.Member;
 import wooteco.subway.service.member.MemberService;
+import wooteco.subway.web.member.InvalidAuthenticationException;
+import wooteco.subway.web.member.LoginMember;
 
 import static org.springframework.web.context.request.RequestAttributes.SCOPE_REQUEST;
