+++ src/main/java/wooteco/subway/web/member/login/WebMvcConfig.java
@@ -1,0 +1,28 @@
+package wooteco.subway.web.member.login;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+    private final BearerAuthInterceptor bearerAuthInterceptor;
+    private final LoginMemberMethodArgumentResolver loginMemberArgumentResolver;
+
+    public WebMvcConfig(BearerAuthInterceptor bearerAuthInterceptor, LoginMemberMethodArgumentResolver loginMemberArgumentResolver) {
+        this.bearerAuthInterceptor = bearerAuthInterceptor;
+        this.loginMemberArgumentResolver = loginMemberArgumentResolver;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(bearerAuthInterceptor).addPathPatterns("/me/bearer");
+    }
+
+    @Override
+    public void addArgumentResolvers(List argumentResolvers) {
+        argumentResolvers.add(loginMemberArgumentResolver);
+    }
+}
