+++ src/test/java/wooteco/subway/AuthAcceptanceTest.java
@@ -13,31 +13,7 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class AuthAcceptanceTest extends AcceptanceTest {
-    @DisplayName("Basic Auth")
-    @Test
-    void myInfoWithBasicAuth() {
-        createMember(TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
-
-        MemberResponse memberResponse = myInfoWithBasicAuth(TEST_USER_EMAIL, TEST_USER_PASSWORD);
-
-        assertThat(memberResponse.getId()).isNotNull();
-        assertThat(memberResponse.getEmail()).isEqualTo(TEST_USER_EMAIL);
-        assertThat(memberResponse.getName()).isEqualTo(TEST_USER_NAME);
-    }
-
-    @DisplayName("Session")
-    @Test
-    void myInfoWithSession() {
-        createMember(TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
-
-        MemberResponse memberResponse = myInfoWithSession(TEST_USER_EMAIL, TEST_USER_PASSWORD);
-
-        assertThat(memberResponse.getId()).isNotNull();
-        assertThat(memberResponse.getEmail()).isEqualTo(TEST_USER_EMAIL);
-        assertThat(memberResponse.getName()).isEqualTo(TEST_USER_NAME);
-    }
-
-    @DisplayName("Bearer Auth")
+    @DisplayName("로그인 후 발급받은 토큰으로 본인의 정보 접근")
     @Test
     void myInfoWithBearerAuth() {
         createMember(TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
@@ -49,19 +25,17 @@ public class AuthAcceptanceTest extends AcceptanceTest {
         assertThat(memberResponse.getName()).isEqualTo(TEST_USER_NAME);
     }
 
-    public MemberResponse myInfoWithBasicAuth(String email, String password) {
-        // TODO: basic auth를 활용하여 /me/basic 요청하여 내 정보 조회
-        return null;
-    }
-
-    public MemberResponse myInfoWithSession(String email, String password) {
-        // TODO: form auth를 활용하여 /me/session 요청하여 내 정보 조회
-        return null;
-    }
-
     public MemberResponse myInfoWithBearerAuth(TokenResponse tokenResponse) {
-        // TODO: oauth2 auth(bearer)를 활용하여 /me/bearer 요청하여 내 정보 조회
-        return null;
+        return
+                given().
+                        auth().oauth2(tokenResponse.getAccessToken()).
+                        accept(MediaType.APPLICATION_JSON_VALUE).
+                when().
+                        get("/me/bearer").
+                then().
+                        log().all().
+                        statusCode(HttpStatus.OK.value()).
+                        extract().as(MemberResponse.class);
     }
 
     public TokenResponse login(String email, String password) {
@@ -74,9 +48,9 @@ public class AuthAcceptanceTest extends AcceptanceTest {
                         body(params).
                         contentType(MediaType.APPLICATION_JSON_VALUE).
                         accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
+                        when().
                         post("/oauth/token").
-                then().
+                        then().
                         log().all().
                         statusCode(HttpStatus.OK.value()).
                         extract().as(TokenResponse.class);
