+++ src/test/java/wooteco/subway/acceptance/line/WholeSubwayAcceptanceTest.java
@@ -1,5 +1,6 @@
 package wooteco.subway.acceptance.line;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.subway.service.line.dto.LineDetailResponse;
@@ -11,6 +12,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@Disabled
 public class WholeSubwayAcceptanceTest extends AcceptanceTest {
     @DisplayName("지하철 노선도 전체 정보 조회")
     @Test
