+++ src/test/java/wooteco/subway/acceptance/path/PathAcceptanceTest.java
@@ -1,6 +1,7 @@
 package wooteco.subway.acceptance.path;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.subway.AcceptanceTest;
@@ -10,7 +11,7 @@ import wooteco.subway.service.station.dto.StationResponse;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
+@Disabled
 public class PathAcceptanceTest extends AcceptanceTest {
     @Override
     @BeforeEach
