+++ src/test/java/wooteco/subway/domain/line/LineRepositoryTest.java
@@ -1,5 +1,6 @@
 package wooteco.subway.domain.line;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
@@ -7,7 +8,7 @@ import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
 import java.time.LocalTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
+@Disabled
 @DataJdbcTest
 public class LineRepositoryTest {
     @Autowired
