+++ src/test/java/wooteco/subway/service/station/StationServiceTest.java
@@ -1,5 +1,6 @@
 package wooteco.subway.service.station;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -14,7 +15,7 @@ import java.time.LocalTime;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
+@Disabled
 @SpringBootTest
 @Sql("/truncate.sql")
 public class StationServiceTest {
